+++ src/neuroconv/nwbconverter.py
@@ -255,7 +255,7 @@ class NWBConverter:
         self.validate_metadata(metadata=metadata, append_mode=append_mode)
         self.validate_conversion_options(conversion_options=conversion_options)
 
-        self.temporally_align_data_interfaces()
+        self.temporally_align_data_interfaces(metadata=metadata, conversion_options=conversion_options)
 
         with make_or_load_nwbfile(
             nwbfile_path=nwbfile_path,
@@ -273,7 +273,9 @@ class NWBConverter:
 
             configure_backend(nwbfile=nwbfile_out, backend_configuration=backend_configuration)
 
-    def temporally_align_data_interfaces(self):
+    def temporally_align_data_interfaces(
+        self, metadata: Optional[dict] = None, conversion_options: Optional[dict] = None
+    ):
         """Override this method to implement custom alignment."""
         pass